[FEAT] 새로고침 시 즐겨찾기 아이콘 유지
